+++ src/platforms/android/snapshot-helper-capture.ts
@@ -1,5 +1,10 @@
 import { AppError } from '../../utils/errors.ts';
 import type { SnapshotOptions } from '../../utils/snapshot.ts';
+import {
+  parseInstrumentationRecords,
+  readInstrumentationResultBoolean,
+  readInstrumentationResultNumber,
+} from './instrumentation-helper.ts';
 import { parseUiHierarchy } from './ui-hierarchy.ts';
 import {
   ANDROID_SNAPSHOT_HELPER_COMMAND_OVERHEAD_MS,
@@ -20,13 +25,6 @@ type AndroidSnapshotHelperChunk = {
   index: number | undefined;
   count: number | undefined;
   payloadBase64: string;
-};
-
-type AndroidInstrumentationRecordState = {
-  status: Array<Record<string, string>>;
-  results: Array<Record<string, string>>;
-  currentStatus: Record<string, string> | null;
-  currentResult: Record<string, string> | null;
 };
 
 export type AndroidSnapshotHelperResolvedCaptureOptions = {
@@ -429,101 +427,10 @@ function readOptionalCaptureMode(
   return value === 'interactive-windows' || value === 'active-window' ? value : undefined;
 }
 
-function parseInstrumentationRecords(output: string): {
-  status: Array<Record<string, string>>;
-  results: Array<Record<string, string>>;
-} {
-  const state: AndroidInstrumentationRecordState = {
-    status: [],
-    results: [],
-    currentStatus: null,
-    currentResult: null,
-  };
+export {
+  readInstrumentationResultNumber as readAndroidSnapshotHelperMetadataNumber,
+  readInstrumentationResultBoolean as readAndroidSnapshotHelperMetadataBoolean,
+};
 
-  for (const line of output.split(/\r?\n/)) {
-    readInstrumentationRecordLine(line, state);
-  }
-  flushInstrumentationRecords(state);
-  return { status: state.status, results: state.results };
-}
-
-function readInstrumentationRecordLine(
-  line: string,
-  state: AndroidInstrumentationRecordState,
-): void {
-  if (line.startsWith('INSTRUMENTATION_STATUS: ')) {
-    state.currentStatus ??= {};
-    readKeyValue(line.slice('INSTRUMENTATION_STATUS: '.length), state.currentStatus);
-    return;
-  }
-  if (line.startsWith('INSTRUMENTATION_STATUS_CODE: ')) {
-    flushStatusRecord(state);
-    return;
-  }
-  if (line.startsWith('INSTRUMENTATION_RESULT: ')) {
-    state.currentResult ??= {};
-    readKeyValue(line.slice('INSTRUMENTATION_RESULT: '.length), state.currentResult);
-    return;
-  }
-  if (line.startsWith('INSTRUMENTATION_CODE: ')) {
-    flushResultRecord(state);
-  }
-}
-
-function flushInstrumentationRecords(state: AndroidInstrumentationRecordState): void {
-  flushStatusRecord(state);
-  flushResultRecord(state);
-}
-
-function flushStatusRecord(state: {
-  status: Array<Record<string, string>>;
-  currentStatus: Record<string, string> | null;
-}): void {
-  if (state.currentStatus) {
-    state.status.push(state.currentStatus);
-    state.currentStatus = null;
-  }
-}
-
-function flushResultRecord(state: {
-  results: Array<Record<string, string>>;
-  currentResult: Record<string, string> | null;
-}): void {
-  if (state.currentResult) {
-    state.results.push(state.currentResult);
-    state.currentResult = null;
-  }
-}
-
-function readKeyValue(line: string, target: Record<string, string>): void {
-  const separator = line.indexOf('=');
-  if (separator < 0) {
-    return;
-  }
-  target[line.slice(0, separator)] = line.slice(separator + 1);
-}
-
-export function readAndroidSnapshotHelperMetadataNumber(
-  value: string | undefined,
-): number | undefined {
-  if (value === undefined) {
-    return undefined;
-  }
-  const parsed = Number(value);
-  return Number.isFinite(parsed) ? parsed : undefined;
-}
-
-export function readAndroidSnapshotHelperMetadataBoolean(
-  value: string | undefined,
-): boolean | undefined {
-  if (value === 'true') {
-    return true;
-  }
-  if (value === 'false') {
-    return false;
-  }
-  return undefined;
-}
-
-const readOptionalNumber = readAndroidSnapshotHelperMetadataNumber;
-const readOptionalBoolean = readAndroidSnapshotHelperMetadataBoolean;
+const readOptionalNumber = readInstrumentationResultNumber;
+const readOptionalBoolean = readInstrumentationResultBoolean;
